feat: switched to codeclimate (superset of gitlab format)
